[chore] fixed tests on CI with storybook
