+++ package.json
@@ -9,6 +9,8 @@
     "install-n-test": "pnpm install && pnpm test",
     "install-n-browser-test": "pnpm install && pnpm browser-test",
     "install-n-axe-test": "pnpm install && pnpm axe-test",
+    "install-n-storybook-test": "pnpm install && npx concurrently -k -s first \"pnpm storybook:build --quiet && npx http-server storybook-static --port 6006 --silent\" \"wait-on tcp:6006 && pnpm run storybook:test\"",
+    "install-n-storybook-snapshot-test": "pnpm install && pnpm storybook-snapshot:test",
     "test": "vitest",
     "test:setup": "docker build -t intergalactic-playwright -f playwright.dockerfile .",
     "test:docker": "pnpm run-in-docker pnpm install-n-test",
@@ -16,11 +18,13 @@
     "browser-test:docker": "pnpm run-in-docker pnpm install-n-browser-test",
     "axe-test": "pnpm playwright test --config playwright.axe.config.ts",
     "axe-test:docker": "pnpm run-in-docker pnpm install-n-axe-test",
+    "storybook-test:docker": "pnpm run-in-docker pnpm install-n-storybook-test",
+    "storybook-snapshot-test:docker": "pnpm run-in-docker pnpm install-n-storybook-snapshot-test",
     "test:types": "vitest typecheck",
     "vo-test:setup": "npx @guidepup/setup",
     "vo-test": "playwright test --config playwright.vo.config.ts",
     "vo-test:ci": "pnpm tsm --require=./.ci/tsm-filter-warnings.js ./.ci/vo-tests-ci.ts",
-    "run-in-docker": "docker run -e PLAYWRIGHT_ARGS= --rm --network host -v $(pwd)/package.json:/work/package.json -v $(pwd)/playwright.browser.config.ts:/work/playwright.browser.config.ts -v $(pwd)/playwright.axe.config.ts:/work/playwright.axe.config.ts -v $(pwd)/allure-results:/work/allure-results/ -v $(pwd)/semcore:/work/semcore/ -v $(pwd)/tools:/work/tools/ -v $(pwd)/website:/work/website/ -v $(pwd)/test-results:/work/test-results/ -w /work/ intergalactic-playwright",
+    "run-in-docker": "docker run -e PLAYWRIGHT_ARGS= --rm --network host -v $(pwd)/package.json:/work/package.json -v $(pwd)/vite.config.ts:/work/vite.config.ts -v $(pwd)/playwright.browser.config.ts:/work/playwright.browser.config.ts -v $(pwd)/playwright.axe.config.ts:/work/playwright.axe.config.ts -v $(pwd)/playwright.storybook.config.ts:/work/playwright.storybook.config.ts -v $(pwd)/playwright.storybook-snapshot.config.ts:/work/playwright.storybook-snapshot.config.ts -v $(pwd)/.storybook:/work/.storybook -v $(pwd)/allure-results:/work/allure-results/ -v $(pwd)/semcore:/work/semcore/ -v $(pwd)/tools:/work/tools/ -v $(pwd)/website:/work/website/ -v $(pwd)/test-results:/work/test-results/ -w /work/ intergalactic-playwright",
     "website": "pnpm --dir website start",
     "lint": "pnpm lint:es && pnpm lint-changelogs && pnpm lint:ts",
     "lint-changelogs": "pnpm --filter @semcore/changelog-handler run lint-changelogs",
@@ -43,8 +47,9 @@
     "i18n:download": "pnpm crowdin download && tsm --require=./.ci/tsm-filter-warnings.js .ci/i18n-csv-2-json.ts",
     "git:setup": "git config --local attributesfile.name \".gitattributes\" && git config --local rebase.backend \"merge\" && git config --local merge.changelog-merge-driver.name \"changelog-merge-driver\" && git config --local merge.changelog-merge-driver.driver \"pnpm tsm --require=./.ci/tsm-filter-warnings.js ./.ci/changelog-merge-driver.ts %A %B\"",
     "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build",
-    "storybook-test": "playwright test --config=playwright.storybook.config.ts"
+    "storybook:build": "storybook build",
+    "storybook:test": "playwright test --config=playwright.storybook.config.ts",
+    "storybook:snapshot-test": "playwright test --config=playwright.storybook-snapshot.config.ts"
   },
   "private": true,
   "devDependencies": {
@@ -68,10 +73,10 @@
     "@semcore/gcs-upload": "workspace:*",
     "@semcore/slack-integration": "workspace:*",
     "@semcore/theme-redesign": "1.5.12",
+    "@storybook/addon-a11y": "8.3.5",
     "@storybook/addon-essentials": "8.3.5",
     "@storybook/addon-links": "8.3.5",
     "@storybook/addon-storysource": "8.3.5",
-    "@storybook/addon-a11y": "8.3.5",
     "@storybook/blocks": "8.3.5",
     "@storybook/react": "8.3.5",
     "@storybook/react-vite": "8.3.5",
@@ -85,6 +90,7 @@
     "allure-js-commons": "3.0.0-beta.6",
     "allure-playwright": "3.0.0-beta.6",
     "allure-vitest": "3.0.0-beta.6",
+    "concurrently": "^9.0.1",
     "csv": "6.2.0",
     "dotenv": "8.2.0",
     "esbuild": "0.14.29",
@@ -92,6 +98,7 @@
     "fast-glob": "3.2.11",
     "fast-hash-code": "2.1.0",
     "fs-extra": "9.1.0",
+    "http-server": "14.1.1",
     "husky": "8.0.0",
     "inquirer": "6.3.1",
     "intergalactic": "workspace:*",
@@ -111,10 +118,11 @@
     "typescript": "5.1.6",
     "typescript-plugin-css-modules": "2.2.0",
     "typescript-workspace-plugin": "2.0.1",
+    "unplugin": "1.14.1",
     "vite": "5.4.9",
     "vite-plugin-babel": "1.1.3",
     "vitest": "1.3.0",
-    "unplugin": "1.14.1",
+    "wait-on": "8.0.1",
     "webdriverio": "8.24.3",
     "yaml": "2.1.1"
   },
