+++ playwright.storybook-snapshot.config.ts
@@ -1,0 +1,8 @@
+import defaultConfig from './playwright.storybook.config';
+
+const config = {
+  ...defaultConfig,
+  testMatch: /\.storybook-snapshot-test.ts(x){0,1}$/,
+}
+
+export default config;
